refactoring test cases and move them to unit test folder
