+++ frontend/src/__tests__/unit/components/TheFAQList.spec.ts
@@ -1,9 +1,9 @@
 import { beforeEach, describe, expect, test, vi } from 'vitest'
-import TheFAQList from '../../components/TheFAQList.vue'
-import { mockFAQs } from '../../__mocks__/stores/mockFAQs'
+import TheFAQList from '../../../components/TheFAQList.vue'
+import { mockFAQs } from '../../../__mocks__/stores/mockFAQs'
 import { createPinia, setActivePinia } from 'pinia'
 import {mount} from '@vue/test-utils'
-import { useFAQStore } from '../../stores/useFAQStore'
+import { useFAQStore } from '../../../stores/useFAQStore'
 import {createVuetify} from 'vuetify'
 import * as components from 'vuetify/components'
 import * as directives from 'vuetify/directives'
